g2 -> g5
